+++ traffic_ops/v3-client/crconfig.go
@@ -56,6 +56,21 @@ func (to *Session) SnapshotCRConfigWithHdr(cdn string, header http.Header) (ReqI
 	return reqInf, err
 }
 
+// GetCRConfigNew returns the raw JSON bytes of the latest CRConfig from Traffic Ops, and whether the bytes were from the client's internal cache.
+func (to *Session) GetCRConfigNew(cdn string) ([]byte, ReqInf, error) {
+	uri := apiBase + `/cdns/` + cdn + `/snapshot/new`
+	bts, reqInf, err := to.getBytesWithTTL(uri, tmPollingInterval)
+	if err != nil {
+		return nil, reqInf, err
+	}
+
+	resp := OuterResponse{}
+	if err := json.Unmarshal(bts, &resp); err != nil {
+		return nil, reqInf, err
+	}
+	return []byte(resp.Response), reqInf, nil
+}
+
 // SnapshotCRConfig snapshots a CDN by name.
 // Deprecated: SnapshotCRConfig will be removed in 6.0. Use SnapshotCRConfigWithHdr.
 func (to *Session) SnapshotCRConfig(cdn string) (ReqInf, error) {